Prépare la mise en prod de la branche indeps (avec la feature cachée)
