+++ netlify.toml
@@ -29,7 +29,7 @@ from = "https://embauche.beta.gouv.fr/règle/contrat-salarié--CDD--motif--compl
 to = "https://embauche.beta.gouv.fr/documentation/contrat-salarié/CDD/motif/complément-formation"
 status = 302
 
-# InFrance PRODUCTION settings 
+# InFrance PRODUCTION settings
 
 [[redirects]]
 from = "https://www.mycompanyinfrance.fr/*"
@@ -130,6 +130,14 @@ from = "https://embauche.beta.gouv.fr/robots.txt"
 to = "/robots.embauche.txt"
 status = 200
 
+# When in iframe, we don't prerender the page
+[[redirects]]
+from = "https://embauche.beta.gouv.fr/"
+query = {iframe=":iframe", s=":e", couleur=":couleur", integratorUrl=":integratorUrl", lang=":lang"}
+to = "/embauche.html?iframe=:iframe&s=:e&couleur=:couleur&integratorUrl=:integratorUrl&lang=:lang"
+status = 200
+
+# Otherwise, we redirect to a prerendered version for the home
 [[redirects]]
 from = "https://embauche.beta.gouv.fr/"
 to = "/prerender/embauche/index.html"
@@ -160,11 +168,6 @@ query = {s = "m"}
 to = "/prerender/infrance/index.html"
 status = 200
 
-[[redirects]]
-from = "/social-security"
-query = {s = "m"}
-to = "/prerender/infrance/social-security/index.html"
-status = 200
 
 [[redirects]]
 from = "/*"
@@ -177,12 +180,6 @@ status = 200
 from = "/"
 to = "/prerender/mon-entreprise/index.html"
 status = 200
-
-[[redirects]]
-from = "/sécurité-sociale"
-to = "/prerender/mon-entreprise/sécurité-sociale/index.html"
-status = 200
-
 [[redirects]]
 from = "/*"
 to = "/mon-entreprise.html"
