+++ source/components/conversation/Conversation.js
@@ -1,17 +1,14 @@
-import Scroll from 'Components/utils/Scroll'
-import { getInputComponent } from 'Engine/generateQuestions'
-import { compose } from 'ramda'
-import React, { Component } from 'react'
-import { withNamespaces } from 'react-i18next'
-import { connect } from 'react-redux'
-import { reduxForm } from 'redux-form'
-import {
-	currentQuestionSelector,
-	flatRulesSelector
-} from 'Selectors/analyseSelectors'
-import * as Animate from 'Ui/animate'
-import Aide from '../Aide'
-import './conversation.css'
+import Scroll from 'Components/utils/Scroll';
+import { getInputComponent } from 'Engine/generateQuestions';
+import { compose } from 'ramda';
+import React, { Component } from 'react';
+import { withNamespaces } from 'react-i18next';
+import { connect } from 'react-redux';
+import { reduxForm } from 'redux-form';
+import { currentQuestionSelector, flatRulesSelector } from 'Selectors/analyseSelectors';
+import * as Animate from 'Ui/animate';
+import Aide from '../Aide';
+import './conversation.css';
 
 export default compose(
 	reduxForm({
@@ -21,7 +18,6 @@ export default compose(
 	withNamespaces(),
 	connect(state => ({
 		conversationStarted: state.conversationStarted,
-		themeColours: state.themeColours,
 		flatRules: flatRulesSelector(state),
 		currentQuestion: currentQuestionSelector(state)
 	}))
